scripts/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
transformers==4.35.0
torch>=1.13.0
librosa==0.10.0
soundfile==0.12.1
openai-whisper==20230314
python-multipart==0.0.6
jiwer==2.6.0
